Move `MinerSlice` to `schemas/`
