+++ src/boltzmann_experimentation/schemas/__init__.py
@@ -1,0 +1,4 @@
+from .miner_slice import MinerSlice
+from .training_components import TrainingComponents
+
+__all__ = ["MinerSlice", "TrainingComponents"]
